requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
huggingface-hub==0.19.4
PyMySQL==1.1.0 